+++ components/Studyroom/CoreTimeSet.tsx
@@ -6,7 +6,7 @@ import { styled } from 'styled-components';
 
 import { getCoretimeID } from '@/apis/coretimes';
 import { getCoretimeList } from '@/apis/studyroom';
-import useModal from '@/hooks/useModal';
+import { useModal } from '@/hooks/Common';
 import { IcPlusBtn } from '@/public/assets/icons';
 import { roomIdState } from '@/recoil/atom';
 import { CoreTimeInfo } from '@/types/studyroom';
@@ -144,8 +144,7 @@ const StCoreBtn = styled.button`
 
   background: ${({ theme, disabled }) =>
     disabled ? theme.colors.Gray3 : theme.colors.Purple3};
-  box-shadow:
-    2.47864px 4.33762px 3.71796px 1.23932px rgba(0, 0, 0, 0.15),
+  box-shadow: 2.47864px 4.33762px 3.71796px 1.23932px rgba(0, 0, 0, 0.15),
     0.61966px 1.23932px 7.43592px 4.33762px rgba(153, 153, 153, 0.3) inset,
     0.61966px 1.23932px 8.67524px 4.33762px rgba(255, 255, 255, 0.15) inset;
