+++ components/Studyroom/CoretimeCard.tsx
@@ -1,6 +1,6 @@
 import { styled } from 'styled-components';
 
-import useModal from '@/hooks/useModal';
+import { useModal } from '@/hooks/Common';
 import { CoreTimeInfo } from '@/types/studyroom';
 import { changeDateFormat } from '@/utils/changeDateFormat';
