+++ components/Studyroom/Issue.tsx
@@ -3,7 +3,7 @@ import { useRecoilValue } from 'recoil';
 import { styled } from 'styled-components';
 
 import { getIssueList } from '@/apis/studyroom';
-import useModal from '@/hooks/useModal';
+import { useModal } from '@/hooks/Common';
 import { IcPlusBtn } from '@/public/assets/icons';
 import { roomIdState } from '@/recoil/atom';
 import { IssueInfo } from '@/types/studyroom';
