+++ containers/Home/HomeContainer.tsx
@@ -1,14 +1,14 @@
-import "firebase/messaging";
+import 'firebase/messaging';
 
-import firebase from "firebase/app";
-import { useEffect } from "react";
-import { useRecoilState, useRecoilValue } from "recoil";
-import { styled } from "styled-components";
+import firebase from 'firebase/app';
+import { useEffect } from 'react';
+import { useRecoilState, useRecoilValue } from 'recoil';
+import { styled } from 'styled-components';
 
-import { createToken } from "@/apis/alarm";
-import { Home } from "@/components/Home";
-import { useFirebaseInit } from "@/hooks/FCM";
-import { fcmTokenState, memberIdState } from "@/recoil/atom";
+import { createToken } from '@/apis/alarm';
+import { Home } from '@/components/Home';
+import { useFirebaseInit } from '@/hooks/FCM';
+import { fcmTokenState, memberIdState } from '@/recoil/atom';
 
 const HomeContainer = () => {
   const memberId = useRecoilValue(memberIdState);
