+++ containers/WebRTC/WebRTCContainer.tsx
@@ -4,6 +4,7 @@ import { useEffect, useState } from 'react';
 import { useRecoilValue } from 'recoil';
 
 import { getCoreEndtime } from '@/apis/coretimes';
+import { useModal, useToggle } from '@/hooks/Common';
 import { useFCMPushAlarm } from '@/hooks/FCM';
 import {
   useChatHandler,
@@ -11,8 +12,6 @@ import {
   useVideoSelector,
   useWebRTC,
 } from '@/hooks/Socket';
-import useModal from '@/hooks/useModal';
-import useToggle from '@/hooks/useToggle';
 import { captureTimeState, memberIdState } from '@/recoil/atom';
 import { getNickName } from '@/utils/getNicknames';
