+++ hooks/Common/useModal.ts
@@ -1,5 +1,5 @@
 // 커스텀 훅 저장하는 폴더
-import { useState } from "react";
+import { useState } from 'react';
 
 export type UseModalReturnType = {
   isShowing: boolean;
