+++ recoil/selector.ts
@@ -1,6 +1,6 @@
-import { selector } from "recoil";
+import { selector } from 'recoil';
 
-import { moonScoreState } from "./atom";
+import { moonScoreState } from './atom';
 
 export const totalMoonScoreSelector = selector<number>({
   key: 'totalMoonScore',
